Refactor 'types' folder with 'models' & 'resolvers'
