+++ src/types/index.ts
@@ -1,6 +1,2 @@
-export * from './AuthPayload';
-export * from './Mutation';
-export * from './Post';
-export * from './Query';
-export * from './User';
-export * from './Scalar';
+export * from './models';
+export * from './resolvers';
