+++ src/types/models/Scalar.ts
