+++ src/types/models/index.ts
@@ -1,0 +1,4 @@
+export * from './AuthPayload';
+export * from './Post';
+export * from './User';
+export * from './Scalar';
