+++ src/types/resolvers/Query.ts
@@ -1,6 +1,6 @@
 import { intArg, queryType, stringArg } from '@nexus/schema';
 
-import { getUserId } from '../utils';
+import { getUserId } from '../../utils';
 
 export const Query = queryType({
   definition(t) {
