+++ src/types/resolvers/index.ts
@@ -1,0 +1,2 @@
+export * from './Mutation';
+export * from './Query';
